pyproject.toml
[project]
name = "novela-scraper-translator"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0,<3",
    "pyyaml>=6.0",
    "google-genai>=1.0",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "cloudscraper>=1.2",
    "tqdm>=4.64",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
export = ["fpdf2>=2.7", "ebooklib>=0.18"]
dev = ["ruff>=0.4", "mypy>=1.9", "pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = false
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
